Bump `go-pinecone` to `v4.0.1` (#22)
## Problem
The CLI has not been updated in quite some time. The Go SDK has seen
several major releases since then. We need to bring `go-pinecone` up to
the current v4 release line.

## Solution
This is just some initial housekeeping / clean up work before we get the
ball rolling on other things in subsequent PRs.

- Bump `go-pinecone` to `v4.0.1`, update related code paths to handle
data model changes
- Remove extraneous print statements - these were left in for debugging
I believe. Printing should generally be handled with the `pcio`
package/wrapper.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [X] None of the above: Updating dependencies

## Test Plan
In order to test, the easiest way is probably to clone the
`pinecone-io/cli` repo, install golang and `goreleaser`, and run things
from the command line. It should be fairly easy to get set up. The
`CONTRIBUTING.md` file has a good rundown of how to get going:
https://github.com/pinecone-io/cli/blob/main/CONTRIBUTING.md

Note: If there's anything that no longer works or is off in
documentation let me know - we're spending a lot of time improving the
CLI and this repo generally.

When testing, set an API key and make sure you can perform index
operations (create_pod, create_serverless, list, describe, delete). The
`pinecone login` flow is still in a bit of a broken state - that'll be
addressed in an upcoming PR.

```bash
./dist/pinecone_darwin_arm64/pinecone config set-api-key <YOUR_API_KEY>

./dist/pinecone_darwin_arm64/pinecone index list
./dist/pinecone_darwin_arm64/pinecone index describe --name <YOUR_INDEX_NAME>

./dist/pinecone_darwin_arm64/pinecone index create-pod -n "test-cli-create-pod" -d 3 -m "cosine" -e "us-east-1-aws" -t "p1.x1" --deletion_protection "enabled"
./dist/pinecone_darwin_arm64/pinecone index create-serverless -n "test-cli-create-serverless" -d 3 -m "cosine" -c "aws" -r "us-west-2"

# etc
```