Refactor oauth2 to use authorization code flow rather than device auth (#24)
## Problem
The CLI currently uses the device authorization flow to authorize with
our auth0 tenant, and retrieve an access token when using `pinecone
login` or `pinecone target`. This auth flow has limitations, and is
ideally meant for input constrained devices such as smart TVs. In order
to target organizations and handle SSO when logging into organizations
that enforce SSO, we need the ability to be more flexible with what
we're passing to the auth servers. Currently, we are limited in that we
cannot easily pass `orgId` through as a URL parameter when
`oauth/device/code` is called:
https://auth0.com/docs/quickstart/native/device/01-login#limitations

After getting a lot of great feedback from @retwere, we feel like the
best thing to do here would be to migrate to using the [Authorization
Code Flow with Proof Key for Code Exchange
(PKCE)](https://auth0.com/docs/get-started/authentication-and-authorization-flow/authorization-code-flow-with-pkce).

## Solution
There are two things that need to change to support this in CLI:

- Refactoring how the CLI authenticates using `golang.org/x/oauth2` -
instead of using `oauth2.Config.DeviceAuth` +
`oauth2.Config.DeviceAccessToken`, we need to use
`oauth2.Config.AuthURL` + `oauth2.Config.Exchange`.
- Updating the "Pinecone CLI" [client in the auth0
tenant](https://github.com/pinecone-io/auth0-config/blob/3418054eedb4dd54b21bc52fc70a66ac3ae8af62/tenants/common/tenant.yaml#L325-L356).
This was handled in this PR:
https://github.com/pinecone-io/auth0-config/pull/275

This PR refactors `GetAndSetAccessToken` to use the authorization code
flow rather than device auth. The biggest change is the need to spin up
a server in order to receive the authorization callback from the
browser. We need to do this because we need to parse the code returned
and then call `Exchange`. We can then use the access token as we did
previously. I've encapsulated all of the logic for handling the auth
exchange in the `GetAndSetAccessToken` and `ServeAuthCodeListener`
methods for now, but some of this should probably be broken out later.

Added two new static HTML files under
`/internal/pkg/cli/command/login/assets/redirect_error.html` and
`/internal/pkg/cli/command/login/assets/redirect_success.html`. These
are bundled with the binary at build using the `go:embed` directive in
`login.go`. For now, this is to give some kind of visible indicator on
the redirect page, I'm hoping to improve this later on.

I've wrapped the `oauth2` dependency in an `Authorization` struct to
replace what we called `DeviceAuth`.

Small fix for `IndentJSON` to prevent double printing JSON output.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
From the user perspective, not much has changed with the authentication
flow. You still call `pinecone login` or `pinecone target`, and we ask
you to confirm opening your browser to initiate authenticating.

You can test by pulling this branch down, building with `goreleaser`,
and running from `./dist/`. Until changes are merged to the auth0
production tenant, they are only available in `internal-beta`, so you
will need to call `pinecone config set-environment staging`. If your
account works at https://internal-beta--pinecone-console.netlify.app/
you should be able to authenticate with the CLI.

```
goreleaser build --single-target --snapshot --clean

./dist/pinecone_darwin_arm64/pinecone config set-environment staging
./dist/pinecone_darwin_arm64/pinecone login

# after you login, try running target and then selecting a different org, re-authenticating, and then targeting a project
./dist/pinecone_darwin_arm64/pinecone target
```


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1210690033044780