Refactor `GetAndSetAccessToken`, remove `pinecone` deprecation warning (#31)
## Problem
We don't want to warn when people use `pinecone`, instead just allow
`pinecone` and `pc` as valid entrypoints.

When calling `pc login` or `pc target`, the local authentication server
is not properly listening for the auth response until the user presses
`[Enter]`. This needs to be more robust to allow people to either
copy-paste the URL themselves, or manually hitting `[Enter]` to
authenticate.

## Solution
- Remove deprecation warning from `cmd/pc/main.go`.
- Refactor `GetAndSetAccessToken`: use separate goroutines to handle
listening for the auth code response, and listening for the user to hit
`[Enter]` through stdin. In both cases, the CLI should receive the code
and continue authentication as expected.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Call `pc` or `pinecone` to work with the CLI - validate there is no
warning displayed in either case.
Call `pc login` or `pc target`, and make sure you can either use
`[Enter]` when prompted, cmd-clicking the URL, or manually copy-pasting
the URL into your browser. In each case, the result should be the same,
and you should see the org and project you're targeting:

```
pc login
Visit https://login.pinecone.io/oauth/authorize?.... to authorize the CLI.

Press [Enter] to open the browser, or manually paste the URL above.
[SUCCESS] Logged in as austin.d@pinecone.io. Defaulted to organization ID: -ORG_ID

[INFO] Target org set to pinecone-official.
[INFO] Target project set pinecone-proj.

Hint: Run pc target to change the target context.

Now try pc index -h to learn about index operations.
```