Add `--project-id` and `--organization-id` flags to `target` command (#44)
## Problem
For projects and organizations we use ID values in all commands in the
CLI except for `target`. This command was originally written before
finalizing the public API definitions, and went with an ease-of-use UX
of using either `--proj` (for project name), or `--org` (for
organization name) to target. This is convenient, but could be confusing
given the usage of project and organization IDs elsewhere across the
admin API.

## Solution
Align the `pc target` command around resource IDs rather than names.
I've kept the ability to use names for now. You'll need to pass one or
the other for organization or project.

I also did a bit of cleanup of package level variables in this PR. We
don't need to be exporting the struct defining command options from each
package, so standardize that.

This PR also resolves bugs related to clearing default API key values.
We were storing the key in two different places (`state`, and `secrets`)
and using one or the other for the presentational layer. I've also made
sure when you're clearing service account or default API key
credentials, the auth context will resolve properly.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Test that the `target` command still works as expected:

```bash
$ pc target --proj "my-project"
$ pc target --project-id "id-value"
$ pc target --org "my-org"
$ pc target --organization-id "id-value"

$ pc target --project-id "id-value" --organization-id "id-value"

# mixing and matching should work
$ pc target --proj "proj-name" --organization-id "id-value"

# errors - no flags, both name and ID, etc
```