Fix `--tags` not being applied to integrated indexes, refactor `/index/create.go` for unit tests (#51)
## Problem
@bradumbaugh flagged a bug around the `--tags` argument not being
applied when creating new integrated indexes through the CLI. The
argument itself was not being passed to the Go SDK, which was an easy
fix.

Because of our general lack of unit testing, and the importance of index
creation and the business logic that's applied, we really needed unit
tests for this command.

## Solution
There were a few functional unit tests validating some of the logic
around inferring index type, but ultimately we need more coverage around
the core logic inside of the `pc index create` command.

- Fix `--tags` not being applied to integrated indexes.
- Refactor `/index/create.go` function signatures and add an interface
to allow for dependency injection and unit testing. None of the logic
itself has changed, but I added a new `runCreateIndexWithService`
function and a `CreateIndexService` interface to allow injecting a mock
Go SDK to allow asserting on what gets provided.

## Type of Change
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Test index creation flows for serverless, integrated, and pod indexes.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1211763561664750