Update to `go-pinecone@v5.0.0`, add e2e testing harness (#52)
## Problem

- `go-pinecone v5.0.0` [was
released](https://github.com/pinecone-io/go-pinecone/releases/tag/v5.0.0),
and we need to pull the new version of the dependency into the CLI.
- We're still lacking any kind of integration / end-to-end tests, which
are going to be important going forward as we add features, refactor
existing code, bump dependencies, etc.

## Solution

- Bump to `github.com/pinecone-io/go-pinecone/v5 v5.0.0` and update
relevant imports across the codebase.
- Add a new `test/e2e/` folder which holds all of the test harness code,
along with specific test files.
  -  `test/e2e/helpers/` include testing utilities: 
- `cli.go`: `CLI` (binary test wrapper) for executing commands against
the CLI and inspect / asserting on the output.
    - `config.go`: Configuration for different tests.
- `wait.go`: Simple polling utilities for waiting for things like
indexes to be ready in tests.
- `main_test.go` - Integration test entrypoint, manages binary
resolution, and configuration setup.
- The rest of the top level files represent tests and assertions against
specific commands:
    - `api_key_test.go`
    - `auth_service_account_test.go`
    - `index_serverless_test.go`
    - `project_list_test.go`
    - `target_test.go`
- Look at `test/e2e/README.md` for specific details on environment
variable configuration, and how to run tests locally or against a
specific binary.
- Added `.github/workflows/e2e.yaml` which allows running integration
tests as a part of the existing `ci.yaml` workflow. Update the `ci`
workflow to run on PRs to main.
- Added required environment variables for integration tests to the
`cli` repo.

This harness will grow and change over time. I wanted to put something
together as a somewhat solid base that allows executing against the CLI
binary consistently as things are changing.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan
Unit and integration tests now run as a part of the `ci` workflow, which
should trigger on ever PR and push to `main`.

You can look at the output for the recent run on this PR
[here](https://github.com/pinecone-io/cli/actions/runs/19325052161/job/55274135017):

<img width="832" height="885" alt="Screenshot 2025-11-13 at 4 24 54 AM"
src="https://github.com/user-attachments/assets/57fefeb2-1d93-4bfb-9e8a-5c281f8fc99c"
/>

The output could use some work, etc - but we get some proper test
coverage of core features for now.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1210928986395208